Rework the entire thing to check only one interface
This makes it easier/possible to say "this interface with these IP
addresses", which is actually what I want; the flexibility around
waiting for multiple interfaces is not something I need much, so it
just makes more sense.. and it straightens out the data model
w.r.t. the nixos option, too.